Update CMakeLists.txt to work with VS 2105
This fix is in keeping with the discovery made in the following
upstream commit:
https://github.com/zeromq/zeromq4-1/commit/
279bcb28430bc48366680b41516bd39853697fcb

The idea is that searching for symbols within a library becomes
tricky when the symbols get mangled. So, we search for a symbol
that doesn't mangle (simply to satisfy the requirement of the)
cmake macro. However, printf is a poor choice to search for due
to its variadic nature. We therefore choose 'fopen' to shunt
the function search.